pyproject.toml
[tool.poetry]
name = "vps_manager_django_api"
version = "0.1.0"
description = "Just practicing"
authors = ["MrMika96 <mikamc1996@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-rest-swagger = "^2.2.0"
drf-orjson-renderer = "^1.6.0"
orjson = "^3.8.0"
psycopg2-binary = "^2.9.4"
drf-spectacular = "^0.24.2"
djangorestframework-simplejwt = "^5.2.2"
python-decouple = "^3.6"
django-filter = "^22.1"
Pygments = "^2.16.1"
pygments-pprint-sql = "^0.1.0"
pytest-django = "^4.6.0"
ruff = "^0.1.14"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
